Refactor NavBar and add MobileSidebar component
