+++ src/components/NavBar/NavBar.tsx
@@ -2,20 +2,10 @@
 
 import Image from 'next/image';
 import Link from 'next/link';
-import {
-  SidebarProvider,
-  Sidebar,
-  SidebarContent,
-  SidebarMenu,
-  SidebarMenuItem,
-  SidebarMenuButton,
-  SidebarTrigger,
-} from '@/components/ui/sidebar';
-import { BRIDE_AND_GROOM } from '@/lib/constants';
-import { useIsMobile } from '@/hooks/use-mobile';
+import { SidebarProvider } from '@/components/ui/sidebar';
+import MobileSidebar from '@/components/MobileSidebar/MobileSidebar';
 
 const NavBar = () => {
-  const isMobile = useIsMobile();
 
   const items = [
     { href: 'nossas-historias/', label: 'Nossas Histórias' },
@@ -52,41 +42,10 @@ const NavBar = () => {
               </Link>
             ))}
           </nav>
-
-          <SidebarTrigger className='md:hidden' />
+          <MobileSidebar items={items} />
         </div>
       </main>
 
-      {isMobile && (
-        <Sidebar side='right' className='md:hidden'>
-          <SidebarContent>
-            <div className='flex flex-col items-center gap-2 p-4'>
-              <Link href='/'>
-                <Image
-                  src={'/svg/logoNavBar.svg'}
-                  alt='Logo Casamento, Maria Eduarda e Rafael Omodei'
-                  height={42}
-                  width={42}
-                />
-              </Link>
-              <span className='text-primary font-arapey text-lg text-center'>
-                {BRIDE_AND_GROOM}
-              </span>
-            </div>
-            <SidebarMenu>
-              {items.map(({ href, label }) => (
-                <SidebarMenuItem key={href}>
-                  <SidebarMenuButton asChild>
-                    <Link href={href} className='text-primary'>
-                      {label}
-                    </Link>
-                  </SidebarMenuButton>
-                </SidebarMenuItem>
-              ))}
-            </SidebarMenu>
-          </SidebarContent>
-        </Sidebar>
-      )}
     </SidebarProvider>
   );
 };
